✨ Protocols: Add support for non-collinear calculations
Replaces the `get_starting_magnetization` utility function for the protocols by the
`get_magnetization` one, which now also accepts the `spin_type` input to deal with
both collinear and non-collinear calculations. The new function returns a single
dictionary with three keys:

* `starting_magnetization`
* `angle1`
* `angle2`

Corresponding to the Quantum ESPRESSO inputs. For `spin_type` = `SpinType.COLLINEAR`,
the angles are set to `None`. Since the API of this function is different, we leave
the old `get_starting_magnetization` function and deprecate it.

The function is also able to read magnetic moment from structure of type
`MagneticStructureData`, implemented in `aiida_wannier90_workflows`. Since we do not
want to make this package a dependency, we cannot add tests for this part of the
functionality. The corresponding code is also clearly isolated so it can be removed,
e.g. in favor of the new `StructureData` class from `aiida_atomistic` that would also
host magnetic properties, but is currently not stable yet.

Additionally, the `Pw2wannier90Calculation` has been modified to handle cases where a
single element exhibits multiple different magnetic moments, when we have to name them
with different labels. To address this, when using external projectors, the projector
files are now copied to the working directory and renamed to match the corresponding
labels, ensuring compatibility with the projector requirements in `pw2wannier90.x`.

Co-authored-by: Marnik Bercx <mbercx@gmail.com>