🐛 `PwParser`: set `num_of_atomic_wfc` to `None` when absent from XML
Previously, we always assumed `@num_of_atomic_wfc` is defined when parsing the XML.
However, if the pseudpotential files do not contain information on the pseudo wave 
functions (as is the case for e.g. SG15), this will give an error when parsing.

The error shows up for more recent versions of Quantum ESPRESSO (> 7). Here we use the
`get()` method to deal with the possible absence of the `@num_of_atomic_wfc` key,
and set the corresponding output to `None` in case the information is missing.