🧪 Fix `test_immigrate` tests
In the following commit:

https://github.com/aiidateam/aiida-core/commit/6ab3ed1bd1c9befdebaad6937c876d06b81ccc4a

The unstashing feature was added, adding a default `unstash` key in the
`metadata.options`. This broke the `test_immigrate` tests, since the `metadata`
is checked versus the one given as an input to the `create_builder_from_file` function.

The (patchy) solution is to simply pop the `unstash` key from the `metadata.options`,
similar as we did for the `stash` one. However, perhaps we should avoid having such
default keys in the options, see:

https://github.com/aiidateam/aiida-core/pull/6826#issuecomment-3230353809