🧪 Add `generate_inputs` fixture
We have a whole set of different fixtures for generating inputs for the various
`CalcJob` classes. However, in some cases, one might want to parametrize a test based
on the entry point of the `CalcJob`, and in that case we don't want to have to include
all the various fixtures as input arguments to the test. Instead, we add the
`generate_inputs` fixture, that allows the tests to generate the inputs of any `CalcJob`
based on their entry point.

This will also be useful to reduce the complexity of the tests in the future: if we only
have one fixture for generating the inputs of any process, we can use it consistently,
parametrize more tests and reduce the amount of code in the tests.