🧪 Test unrecoverable failures for `BaseRestartWorkChain`s
The `BaseRestartWorkChain` has a feature that the first time it encounters an unhandled
error, it restarts the process. However, in case the process fails twice consecutively
with an unhandled error, the `BaseRestartWorkChain` is finished with an
`ERROR_SECOND_CONSECUTIVE_UNHANDLED_FAILURE`. This is to deal with transient errors,
for example in the case of node failure.

Here we add a test to verify this behaviour for the `BaseRestartWorkChain`s.