🔧 Update CI/nightly actions to use `uv`
`uv` is a much faster Python package installer/resolver than `pip`, written in Rust,
see:

https://docs.astral.sh/uv/

Since the installation of the dependencies is now reduced to ~3s from ~50s, we can
remove the caching step in the CI and nightly actions.

We also set `FORCE_COLOR` to `1` in the CI/nightly environments to make our logs more
colorful, making them easier to parse.

Co-authored-by: Daniel Hollas <daniel.hollas@bristol.ac.uk>