🗑️ Remove `handle_unrecoverable_failure` error handler
The `handle_unrecoverable_failure` error handler, implemented on all
`BaseRestartWorkChain` classes, takes any failure of the wrapper `CalcJob` class with
an exit status < 400 and aborts the work chain. This overrides the default behaviour
of the `BaseRestartWorkChain` which would normally only abort the work chain after the
second consecutive unhandled failure.

However, this feature of the `BaseRestartWorkChain` is desirable in some cases, such as
when there is a node failure. The `handle_unrecoverable_failure` process handler would
also finish the workflows in these cases when we don't want it to. Hence, we remove it.

We also deprecate the `ERROR_UNRECOVERABLE_FAILURE` exit code for all
`BaseRestartWorkChain` classes that no longer return it at this stage.