🧪 CLI: Use explicit list instead of `recurse_commands`
The `test_commands.py` module uses a neat `recurse_commands` function to use the `click`
root command to recursively build a list of all available commands. This is then used
to test running the command with the `--help/-h` option.

While neat, it seems building the list of commmands with `recurse_commands` affects the
AiiDA profile from the `aiida_profile` test fixture. It's not entirely clear how, but
this affects the logs of the `pw.x` parser tests, and breaks the test suite.

Further isolation reveals that using `get_command(...)` to obtain a certain command,
for example:

```
from aiida_quantumespresso.cli import cmd_root
cmd_root.get_command(
   None, 'workflow'
).get_command(
   None, 'launch'
).get_command(None, 'q2r-base')
```

And _then_ running the corresponding CLI test (`tests/cli/workflows/test_q2r.py`),
breaks the `pw.x` parser tests in `tests/parsers/test_pw.py` which check that the logs
of a failed calculation are not empty, because they are.

Following the KISS principle, we remove the `recurse_commands` function in favor of just
parametrizing the list it generates. For those with a passion for recursion and more
patience than this committer, we leave behind an easter egg for you to play with. 🥚