🧪 Update `PwParser` fixtures to QE 6.6 XML output
The earliest versions of the `PwParser` exclusively parsed from the stdout because that
was all that `pw.x` produced. At some point, it started also writing output in an XML
file, however, this format was deprecated as of QE v6.2 and replaced by an XML with a
different schema. The advantage of the latter is that it comes with an XSD file which
can be used to validate and parse XML output files.

For backwards compatibility reasons, the `PwParser` has never fully transitioned to
parsing exclusively from the XML. This is also because the XML does not provide all
data that was currently parsed from the stdout.

The time has come to start getting rid of the stdout parsing as it is intrinsically
fragile. The current version of `aiida-quantumespresso` anyway only supports QE v6.6
and up, all of which have full support for the schema based XML output. The idea is to
migrate the `PwParser` to parse almost exclusively from the XML file and only fetch that
from the stdout which the XML does not provide.

To prepare this move, the test fixtures, which often contain reference files generated
by old QE versions with the old deprecated XML format, need to be updated to provide
reference output with modern QE versions. Here we replace the output files with output
generated by QE pw.x v6.6 or above.

For the "success" calculations, a new calculation was run and the output files were
used. The same technique was applied for certain "failed" cases wherever that was
possible by setting a specific input. For example, the out-of-walltime error could
easily be simulated by setting the `max_seconds` input to 1 second, guaranteeing the
error to be hit. There are however a number of errors that are very specific and
difficult to reproduce, notably the BFGS errors for the vc-relax tests. For these, the
XML has been replaced with outputs of a similar calculation that actually succeeded.
These files are then naturally not commensurate with the stdout file, but this is not
important since they are testing the parsing of the error, which is only done from the
stdout anyway.

Following are notes related to changes for specific tests cases:

* `vcrelax_success_rVV10`: Has been renamed to `scf_success_rVV10`. The reason being
  that the calculation had to be reperformed and a `vc-relax` kept failing with a
  `ERROR_IONIC_CONVERGENCE_REACHED_FINAL_SCF_FAILED`. Since the relaxation is not the
  point of the test, an `scf` calculation was done instead.

* `test_pw_vcrelax_failed_bfgs_history_already_converged`: we were unable to reproduce
  this case for QE v6.6 or above. In fact the test output seemed to be bogus, there
  was no BFGS reset done at all? In case we run into this issue in the future, we can
  always double-check the logic and reintroduce the test.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>
Co-authored-by: Timo Reents <77727843+t-reents@users.noreply.github.com>