♻️ `PwParser`: primarily parse from XML instead of `stdout`
As is extensively explained in a5637b3abafd88a770ec01687adff05f9ca07505, the parser of
the `pw.x` code of Quantum ESPRESSO was historically focused on parsing from the
`stdout`, since in the begin times there was no XML output to speak of. However, parsing
the `stdout` is often convoluted and fragile, breaking the parser with minor changes in
how the output is printed.

After updating the tests in a5637b3abafd88a770ec01687adff05f9ca07505, we update the
`PwParser` to rely on the XML for the majority of the output. To preserve backwards
compatibility, we still take data missing from the XML for supported Quantum ESPRESSO
versions from the `stdout` using the `backwards_compatibility_trajectory` and
`backwards_compatibility_parameters` methods.

We also add the units for several energy values that were previously missing in the
`output_parameters`.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>
Co-authored-by: Timo Reents <77727843+t-reents@users.noreply.github.com>