🐛 `PdosWorkChain`: Fix `nscf` `parent_folder` usage
The `PdosWorkChain` can either run the SCF in the workflow, or skip that step in case
the user provides a `pw.parent_folder` input in the `nscf` namespace. However, the
second option is currently broken in multiple ways. Here we provide fixes for several
issues:

* the `pw.parent_folder` input is excluded from the `nscf` namespace, so can't be set.
  -> We remove `pw.parent_folder` from the `exclude` tuple.
  
* in case the `nbands_factor` is provided in the inputs, the `run_nscf` step will
  attempt to obtain the number of bands in the SCF from a non-existent `workchain_scf`
  key in the context.
  -> We obtain the number of bands from the `creator` of the `parent_folder` instead.

* in case `serial_clean` is set to true, the remote folders of the `workchain_scf` are
  cleaned in the `inspect_nscf` step. But `workchain_scf` is not stored in the context
  when the SCF is not run.
  -> We only execute this step if both `serial_clean` and `scf` is specified in the
  inputs.

Additionally, we improve the `validate_inputs` docstring to include all warnings/errors
raised.