👌 `NebCalculation`: Use `images` instead of `first/last_structure`
In order to provide more control over the path to the user, we adapt the inputs of the
`NebCalculation` to use a single `images` input with type `TrajectoryData` that
contains all the structures, instead of only the `first_structure` and `last_structure`.
The latter are deprecated and will be removed in the next major release.

We also add the `validate_inputs` top-level validator to make sure the required inputs
are provided and add a deprecation warning. Validation related to these inputs that was
previously checked in the `prepare_for_submission` method is also moved to this
validator so these issues are discovered before submission.

Co-authored-by: Davide Bidoggia <davide.bidoggia@units.it>
Co-authored-by: ionuorah <anyianyi89@gmail.com>