👌 Protocols: set `resources` based on scheduler
The current protocols set `num_machines` in the `resources`, which works fine for
several schedulers like `core.slurm` or `core.direct`. However, this choice fails for
e.g. the `core.sge` scheduler, that doesn't accept the `num_machines` key in the
resources.

Ideally, we'd be able to get reasonable defaults from the scheduler, but this feature
isn't available (yet). As a stopgap solution, we remove the `resources` from the
protocol YAML files. To preserve backwards compatibility, and keep the usage of the
protocols as simple as possible, we add the `set_default_resources` static method to the
`ProtocolMixin`, and use that to set default resources in the
`get_builder_from_protocol()` methods.

Co-authored-by: Timo Reents <77727843+t-reents@users.noreply.github.com>