📚 Update "getting started" path
The documentation on the installation and first `pw.x` tutorial were a bit outdated and
did not focus on the features we want the user to learn first. Here we adapt both
sections, as well as the flow to connect them.

On the landing page, we guide the users to the installation page and the (new) quick
start via two panels.

> Installation

With `verdi presto`, things here can be much simplified. The basic installation now
describes three things:

1. Install the package, no need to discuss install `aiida-core`. Then set up a profile
   with `verdi presto`, which also sets up the `localhost` computer.
2. Set up the basic codes with the new CLI feature introduced in the commit
   8d10960d80e2b09ac45606b0d2df9382bbee16ee
3. Installing the pseudopotentials with `aiida-pseudo`.

The CLI tab-completion section is moved to to an "Advanced" section at the bottom of
the page, as it should really not be a priority for most users. To keep users on
point, we add another panel to guide them to the new "Quick start".

> Getting started

This is the first place where we want users to start learning how to use the package.

1. Instead of starting by using the CLI, we show them how to run a `pw.x` calculation
   with the protocols via the Python API. This is how we _want_ beginners to first use
   the package. Here we still use `run`, to keep things simple and robust.
2. One of the first things users will want to do is adapt the inputs. We show them
   here how to use overrides and adapt the builder. Two more approaches that start
   from scratch are added to a preliminary "How to customise inputs" page.
3. Finally, we show how to `submit` the process instead of run it. For _anything_ real
   they want to run, they have to understand this. By using `submit`, it also becomes
   natural to discuss the daemon, and how to inspect processes.