👌 Fix entry point of `realhydrogen` to `core.realhydrogen`
The `RealhydrogenOrbital` class was still loaded with the old entry point, here add the
`core.` prepend to remove deprecation warnings.

Note that the `spinorbithydrogen` and `noncollinearhydrogen` entry points are defined
in this repository, so they don't need the `core.` prefix. In principle we should add
the `quantumespresso.` prefix, but frankly I think we should move these to
`aiida-atomistic`.