`PwRelaxWorkChain`: Fix meta-convergence
This PR fixes two issues related to the meta-convergence of the
`PwRelaxWorkChain`.

* In the current version, we check whether the maximum number of
  iterations has been reached in the `should_run_relax` method. If yes, we return an exit code. However,
we don't check whether the last iteration was actually successful and
converged with respect to the meta-convergence. To also cover those
cases, the `self.ctx.converged` variable is (re)introduced.

* Related to the previous aspect, we currently return an exit code
  within the `should_run_relax` method in case the maximum number of
meta-convergence iterations was reached. However, this method is used as
the condition of the `While_` loop and the `plumpy` implementation only
supports boolean return values
(https://github.com/aiidateam/plumpy/blob/7df19a1bc5e59c8b9e67616ac9309145966ae867/src/plumpy/workchains.py#L401-L414). Hence, the current version could get
stuck in infinite loops (e.g., the k-mesh keeps changing), as the returned exit code is not respected.
This was fixed by changing the position of this statement and moving it
to the `results` method