🐛 `NebCalculation`: Fix top-level inputs validation
In b0b63c1c2493949a8ffe95fe59ae2da98ffe6b62 we added a top-level validator to the
`NebCalculation`. This approach is solid, and warns/stops the user before runtime if
there are issues with the inputs.

However, in doing so we override the default `CalcJob` validation. To fix this, we
take the same approach as is used in the `PwCalculation`: manually apply the
`validate_calc_job` validation function on the top-level inputs in the
`NebCalculation.validate_inputs` method.

We also change the name of the `inputs` input argument to `value`, to be consistent
with other validators.