🔧 CI: Move integration job to nightly build
The integration run is a nice addition to our CI, but unfortunately the installation
of the SSSP from the Materials Cloud can fail in case it is unavailable, see:

https://github.com/aiidateam/aiida-pseudo/issues/194

This means that the check will fail on open PRs and commits. Admins can still merge,
of course, but the failures lead to worries/noise for contributors, and a deep sense of
unease for developers with a borderline OCD-like desire to see green checkmarks.

Here we move the integration test to the nightly build. This means open PRs will no
longer be integration tested, but we should have enough unit tests to deal with that.
We can always reintroduce the integration test to the CI once it is less fragile.
Having the integration test in the nightly build also means we'll be notified via Slack
about any issues reasonably quickly, even at times when developers are distracted by
other projects.