✨ CLI: Add command for setting up multiple codes
Setting codes can be a challenge for new users, because they have to provide quite a
few inputs correctly, and will not understand why this is necessary. This is the
current command for setting up the `pw.x` code in our installation docs:

  verdi code create core.code.installed -n --computer localhost --label pw \
    --default-calc-job-plugin quantumespresso.pw --filepath-executable pw.x

A few issues:

1. The command from `aiida-core` is confusing.
2. The `-n` is necessary to make the command non-interactive.
3. You have to provide `pw` in three places:
   * `--label`
   * `--default-calc-job-plugin`
   * `--filepath-executable`

Making this easier in `aiida-core` is certainly possible, but tricky because changes
will need to be backwards compatible and will always need to be more flexible to
support many use cases.

Another issue is that the user will typically have to run this command several times,
once for each of the Quantum ESPRESSO binaries (`pw.x`, `projwfc.x`, `dos.x`, ...).

Here we add a new command that allows the user to more easily set up multiple codes
for a computer:

  aiida-quantumespresso setup codes localhost pw.x projwfc.x dos.x

The command tries to find the executables in the `PATH` using `which`. The user can
provide the necessary `--prepend-text` and `--append-text` text via command-line
options, or using an editor by passing the `--interactive` option. Additionally, the
user can also specify the path to the directory in which the executables are stored.
The default label is simply the executable without the extension (pw.x -> pw), but a
`--label-template` can also be used. 

Along with the CLI command, we also provide a Python API with similar features. Here
there is no need for an interactive option, since writing multiline strings is easier.
We do add the `increment_label` option in case the user wants to automatically add
an (incrementing) number suffix (`-N`) to the label.

Co-authored-by: Marnik Bercx <mbercx@gmail.com>