👌 `HubbardUtils`: improve behaviour of `get_intersites_radius`
The `get_intersites_radius` utility method of `HubbardUtils` was returning a radius
that was the smallest to contain only the nearest neighbors. This meant that not
necessary all the neighbors found by the nearest neighbors analysis were within such
radius. This might have happened when other atoms were closer to the NN atoms, but
they were not considered NN (i.e., the NN analysis did not consider them as
"ligands'').

Since this routine is mainly of use in the `aiida-hubbard` plugin, we here change its
behavior to instead return the largest radial cutoff to include all of the atoms, and
warn the user if other atoms other than the ligands are within such radius.