📚 Reorganise How-to sections
The current how-to sections are organised by "calculations" and "workflows", and list
each and every one. This might make sense for the "topic guides", but for the how-to's
this may not be the best approach. Moreover, many of these are still "to be added", and
history has shown we don't have a lot of time to write documentation.

Here we adapt the structure of the How-to sections:

* We take the 5 how-to sections we have currently and place them at the top level of the
  nav bar, making them easier to find.
* We adapt the titles from just the process to a more descriptive one, which is more
  understandable than just the process.
* We remove any pages that are currently still empty. Scaffolding is fine in case you
  have time to fill in the details in the near future, but here it is just distracting
  and reveals that our documentation is quite incomplete.