✨ `projwfc.x`: add support for projections onto boxes and total PDOS
Add support for parsing `ldosboxes` files when user sets `tdosinboxes` to `.true.`.
Since as far as we know the user can only "project" in one of two ways:

1. onto orthogonalized atomic orbitals
2. integrated over volumes (also known as boxes)

We check the `stdout` for the appropriate strings and call the corresponding methods.
In case we cannot detect the projection type, it's possible the stdout was missing or
incomplete. Hence, we raise a warning, but still return the corresponding stdout
exit code.

Additionally, we add support for parsing the total projected density of states (PDOS),
for both approaches.

Co-authored-by: ValkScripter <161381539+ValkScripter@users.noreply.github.com>