👌 `PwParser`: Parse total/absolute magnetization from XML
The `total_magnetization` and `absolute_magnetization` values in the `output_parameters`
are currently obtained from the trajectory parsed from the `stdout`. This means the
precision of these outputs is quite low, and the values are missing when running
SIRIUS-QE, for which the `stdout` parsing of these properties is not supported.

Here we parse both from the XML, and remove them from the
`final_trajectory_frame_to_parameters` method that would override the XML-parsed values
with the `stdout` ones from the trajectory.  The `absolute_magnetization` magnetization
should always be present, even for non-spin-polariser calculations. The
`total_magnetization` is not, but is also obviously zero for the system. For consistency
between both parameters and between the outputs of various magnetic settings we set
their value to zero in case either is missing from the XML.