✨ `matdyn.x`: add DOS support
Add support to the `MatDynCalculation` plugin for setting `dos` to true, and in this
case converting the kpoints input to the `nkX` tags instead of providing them as a list.
For `matdyn.x` calculations where`dos` is true, parse the phonon DOS instead of the
phonon bands, and provide this as an output.