👌 `setup_codes`: warn when executable does not exist with `directory`
When using the `directory` input of the `setup_codes` tool, the command will raise a
`FileNotFoundError` when the constructed `<directory>/<executable>` path does not exist.
This means that in case the user relies on another approach for making the executable
accessible than the prepend text, they are unable to set up their codes with
`setup_codes` and the corresponding new CLI command.

Here we adapt the `get_executable_paths` function to raise a warning instead of an
error when the `<directory>/<executable>` path does not exist.