🐛 `setup_codes`: use computer prepend text
When trying to locale the path to the executables with `which`, the
`get_executable_paths` function was not using the computer prepend text, falsely
assuming that the user would specify all required prepend text in the code setup.

Here we merge the computer and code prepend texts before executing the `which` command.