👌 `PwBands`: improve `nbnd` for non-collinear calculations
The number of bands (`nbnd`) we calculate in the `bands.x` step of the
`PwBandsWorkChain` is determined by multiplying the `number_of_bands` output parameter
of the SCF with the `nbands_factor` input. However, we currently assume there are
two electrons in every band, which is only the case for collinear calculations.

Here we use `noncolin` from the `pw.x` inputs to correct this logic for non-collinear
calculations, where there is only one electron in every band.