🐛 `pw.x`: fix units for XML stress/forces parsing
In fdb03d1cd9618a59da545c4e2eb85e3e09f18361, we switched to parsing as much output from
the XML as possible for the `PwParser`. However, we missed that the units in the XML
are different for forces and stresses, and hence there is not an inconsistency between
the units we claim in the `output_parameters`, and those of the actual values in the
`output_trajectory`.

Here we correctly convert:

* Forces to eV / Angstrom
* Stress to GPascal