🔧 Build: Update build/devops envs to Hatch
Switch to using Hatch as the build system. Hatch also allows us to define environments
and scripts to run for development operations, which removes the need for tox.

We remove `pylint` from the pre-commit for now, since it was causing issues in the CI
and will be replaced by Ruff.

We keep the `docs` optional dependencies instead of just specifiying them in the
corresponding Hatch environment, to keep the current RTD configuration in place.

Interestingly, there were still some blatant syntax errors in the repo related to
f-strings that were not caught previously. These are also fixed here.