🧪 `pw.x`: rename rvv10 test
The rVV10 test had been renamed in cb8a5445aedfc12270556875089090f57cd2d0c7 to adhere
to Ruff linting rules. However, this caused a second pytest regression file to be
generated, so we had both:

* test_pw_scf_success_rVV10.yml (uppercase)
* test_pw_scf_success_rvv10.yml (lowercase)

This was causing collision issues when for git:

  warning: the following paths have collided (e.g. case-sensitive paths
  on a case-insensitive filesystem) and only one from the same
  colliding group is in the working tree:

    'tests/parsers/test_pw/test_pw_scf_success_rVV10.yml'
    'tests/parsers/test_pw/test_pw_scf_success_rvv10.yml'

Here we rename the file to a completely new one to fix this issue.