🔧 Deploy: fix `update_changelog.py` script for support branch
The `update_changelog.py` was comparing the commmits of the latest `origin/main` with
the last release tag, instead of the cherry-picked commits for the release of the
support branch. When doing a support release, typically `HEAD` will have all the commits
we cherry-picked.

We also caught a bug that was introduced while switching linters: the `commit_line` was
not always properly assigned to the `commit` variable while adding each entry to the
changelog.

We also adapted the Ruff rules to make them less restrictive.