🔧 CI: use Hatch for nightly build
In cddf25a97b116769f43c582ac52984ccefe26a42 we switched to Hatch for our build system
and adding environments/scrips for running various devops tasks in separate Python
environments. The regular CI was updated correctly, but the nightly build also relied
on the `tests` optional dependencies that were removed.

Here we adapt the nightly build to also rely on the Hatch environments.