🔧 Hatch: clean up `pre-commit` environment
In cb8a5445aedfc12270556875089090f57cd2d0c7 we decided to move to the Ruff linter,
after removing `pylint` from the pre-commit in cddf25a97b116769f43c582ac52984ccefe26a42.
However, we forgot to remove `pylint` and the AiiDA plugin as dependencies of the
corresponding Hatch environment. This is corrected here.