🧪 Warnings: Update `pytest` and correct tests
In our tests, we incorrectly assumed that `with pytests.warns(None)` verifies that no
warnings are raised in the corresponding context. But that is not the case, see the
changelog for `pytest`: 7.0.0rc1, and issue number 8645:

https://docs.pytest.org/en/stable/changelog.html#id479
https://github.com/pytest-dev/pytest/issues/8645

We ran into this problem while updating `pytest` to `~=8.4`, and corrected the usage
with the `recwarn` fixture, or using `nullcontext` in case a test parametrizes both
cases where warnings are raised or not.