🔧 tools: diversify options for developers
In cddf25a97b116769f43c582ac52984ccefe26a42, we made the switch to Hatch for our build
backend and development environments. In doing so, we removed some development extras,
since they were in principle no longer necessary. However, not all contributors are
familiar with Hatch and/or would like to use it. We now effectively force them to do so,
or deal with manually installing the dependencies required to e.g. run the tests.

Here we diversify options for developers:

1. We reintroduce the original extras that developers were used to, so they can continue
   their workflow uninterrupted for now.
2. We add the `dev` depedency group, which `uv` uses by default in its environments.

See the following page for more information on dependency groups:

https://packaging.python.org/en/latest/specifications/dependency-groups/

The changes here allow users of all three approaches to contribute to the package. To
help users get started, we add a quick start section to the developer documentation.