🐛 `ProjwfcParser`: convert `bands` array into floats
The 'bands' array in the `bands` `BandsData` output (phew) is obtained using a regex
pattern in combination with the `findall` method. However, by default the `findall`
method will simply return the extracted values as strings.

Here we correctly case the 'bands' array data type as float when constructing the
numpy array.

Co-authored-by: Xing Wang <xingwang1991@gmail.com>