🐛 `create_kpoints_from_distance`: respect `pbc`
In the case of a symmetric cell, the `create_kpoints_from_distance` ensures that the
k-points mesh is symmetric as well. This is sensible, but in the current approach the
periodic boundary conditions (pbc) of the input `structure` are not respected, and a
symmetric mesh is returned even if the structure is not periodic along one or more
directions.

Here, we consider the periodic boundary conditions even in the case of symmetric cells
by explicitly setting the mesh to 1 in any non-periodic direction.

Co-authored-by: AndresOrtegaGuerrero <34098967+AndresOrtegaGuerrero@users.noreply.github.com>