✨ `projwfc.x`: add support for `kresolveddos`
Add support for k-point resolved calculations with `projwfc.x`, removing `kresolveddos`
from the `ProjwfcCalculation._blocked_keywords`. When doing k-point resolved
calculations, the output files for PDOS and LDOS have an `ik` column prepended. We
detect this column by reading the header of each file, and reshape the outputs
accordingly. The output PDOS and LDOS now are defined on a two-dimensional
(<kpoints_index>, <energy>) grid, stored in the `energy` output.