📚 `README.md`: trim down and refer to documentation
The current README file shows the entire compatibility matrix, and still refers to
tools for development that we no longer support (e.g. `tox`). Instead of updating the
content, we clean up the README and refer to the documentation.

We also correct a small issue with regard to the Python versions in the compatibility
matrix in the documentation.