📚 `conf.py`: ignore missing `aiida-core` references
The `aiida-core` AutoAPI docs have been (temporarily) disabled in:

https://github.com/aiidateam/aiida-core/pull/7056

Since we still have several intersphinx references to the `aiida-core` AutoAPI docs,
our build started failing on RTD. Here we ignore any issues with missing `aiida-core`
references by adding a broad regex to the `nitpick_ignore_regex`.

We could also remove the references, but we haven't settled yet on whether to
reintroduce the AutoAPI docs in the future (also for this package).