💥 `parameters`: enforce consistent casing
Throughout the codebase, we always face the issue that while FORTRAN namelists and
the parameters specified therein are not case-sensitive, the dictionaries we store the
parameters in are. This means we always have to consider the case when we e.g.:

- Write calculation input files
- Perform input validation
- Execute any logic based on the parameters

This makes our code much more verbose, and error prone (see below). Moreover, allowing
freedom in the case of the input parameters makes it much more difficult to write robust
queries. Simply put: allowing for case-insensitive inputs in the parameters is not a
sustainable approach.

Here, we add validation to the `parameters` input port of the various calculation
classes that makes sure they adhere to the following conventions:

1. Namelist keys should be UPPERCASE, and their values must be a dictionary of
   parameters.
2. Parameter keys should be lowercase.

The new `validate_parameters` validator auto-corrects the inputs and emits warnings in
case the node is not yet stored, and rejects stored nodes with error message.

By adding this validation, we can remove any logic that guards against differences in
case, typically via the `_uppercase_dict`/`_lowercase_dict` functions in the
`prepare_for_submission` methods.

Note that some other logic in the codebase has to be adapted to adhere to these new
conventions. An example here is the `la2F` parameter of the `INPUT `namelist of `q2r.x`.
Here we have to adapt all of the logic to use `la2f` instead. This points to an issue
that was previously not considered: in case the user would have used a different casing,
all of this logic would fail.

Similarly, we have to update the protocol/validation/logic for the `PDosWorkChain` for:

- `DeltaE` -> `deltae`: note that here the QE input manual is inconsistent between codes. It uses
  `DeltaE` for `dos.x`, but `deltaE` for `matdyn.x`. To be fair, the units are also
  different.
- `Emin`/`Emax` → `emin`/`emax`

In all cases, we stick to one convention: lowercase for parameters.