✨ PwTools: add method to parse and return occupation matrices (#1215)
When performing Hubbard-corrected DFT calculations, Quantum ESPRESSO
calculates the occupations matrices related to the target manifolds.
The final matrices are saved in the xml file, which it is always retrieved.

To access these matrices, instead of storing them and duplicate the data,
we here add an auxiliary method to parse these matrices and return them 
in a format compatible with pandas, so that they can be used for post-processing
purposes (e.g., calculation of oxidation states, magnetic analysis, ...).

This has been implemented as a "tool" method, such that one can access
the data as follows:
```python
occupations = load_node(<PwCalculation PK/ID>).tools.get_occupations()
```